requirements.txt
allure-pytest==2.8.13
allure-python-commons==2.8.13
atomicwrites==1.4.0
attrs==19.3.0
colorama==0.4.3
importlib-metadata==1.6.0
more-itertools==8.2.0
packaging==20.3
pluggy==0.13.1
py==1.8.1
pyparsing==2.4.7
pytest==5.4.2
selenium==3.141.0
six==1.14.0
urllib3==1.25.9
wcwidth==0.1.9
zipp==3.1.0
